bug fix and addition of Owns() to buffers.
git-svn-id: https://svn.syssoft.uni-trier.de/svn/prisma/DeltaWorks/include@604 87af6b1e-9b34-e211-997a-bc305bce1b9a

+++ engine/gl/raster_font.cpp
@@ -166,7 +166,7 @@ namespace Engine
 	{
 	    float result(0);
 	    for (index_t i = 0; i < len; i++)
-	        result+= width[(unsigned)*str++];
+	        result+= width[(BYTE)*str++];
 	    return result;
 	}
